PB-1498: Moved git version code in a separate package to allow to reuse it
