+++ k8s-validate/cmd/root.go
@@ -5,6 +5,7 @@ import (
 	"os"
 	"runtime"
 
+	"github.com/geoadmin/tool-golang-bgdi/lib/version"
 	"github.com/spf13/cobra"
 )
 
@@ -34,7 +35,7 @@ var versionCmd = &cobra.Command{
 	Use:   "version",
 	Short: "Print the version",
 	Run: func(_ *cobra.Command, _ []string) {
-		fmt.Println(GetGitVersion())
+		fmt.Println(version.GetGitVersion())
 	},
 }
