+++ lib/version/main.go
@@ -1,4 +1,4 @@
-package cmd
+package version
 
 import (
 	"os/exec"
